+++ smarts/core/smarts.py
@@ -53,6 +53,7 @@ from .sumo_traffic_simulation import SumoTrafficSimulation
 from .traffic_history_provider import TrafficHistoryProvider
 from .trap_manager import TrapManager
 from .utils import pybullet
+from .utils.math import rounder_for_dt
 from .utils.id import Id
 from .utils.pybullet import bullet_client as bc
 from .utils.visdom_client import VisdomClient
@@ -84,6 +85,7 @@ class SMARTS:
         self._envision: EnvisionClient = envision
         self._visdom: VisdomClient = visdom
         self._timestep_sec = timestep_sec
+        self._rounder = rounder_for_dt(timestep_sec)
         self._traffic_sim = traffic_sim
         self._motion_planner_provider = MotionPlannerProvider()
         self._traffic_history_provider = TrafficHistoryProvider()
@@ -113,7 +115,7 @@ class SMARTS:
             ActionSpaceType.LaneWithContinuousSpeed,
             ActionSpaceType.Trajectory,
             ActionSpaceType.MPC,
-            ActionSpaceType.Imitation,
+            # ActionSpaceType.Imitation,
         }
 
         # Set up indices
@@ -230,9 +232,7 @@ class SMARTS:
         extras = dict(scores=scores)
 
         # 8. Advance the simulation clock.
-        # round due to FP precision issues, but need to allow arbitrarily-small dt's
-        dec_digits = len("{}".format(self._timestep_sec)) - 2
-        self._elapsed_sim_time = round(self._elapsed_sim_time + dt, dec_digits)
+        self._elapsed_sim_time = self._rounder(self._elapsed_sim_time + dt)
 
         return observations, rewards, dones, extras
 
@@ -547,9 +547,12 @@ class SMARTS:
                     # This is not a pybullet agent, but it has an avatar in this world
                     # to make it's observations. Update the avatar to match the new
                     # state of this vehicle
+                    # XXX: this needs to be disentangled from pybullet.
                     pybullet_vehicle = self._vehicle_index.vehicle_by_id(vehicle_id)
                     assert isinstance(pybullet_vehicle.chassis, BoxChassis)
-                    pybullet_vehicle.control(pose=vehicle.pose, speed=vehicle.speed)
+                    pybullet_vehicle.control(
+                        pose=vehicle.pose, speed=vehicle.speed, dt=self._timestep_sec
+                    )
             else:
                 # This vehicle is a social vehicle
                 if vehicle_id in self._vehicle_index.social_vehicle_ids():
@@ -557,6 +560,8 @@ class SMARTS:
                 else:
                     # It is a new social vehicle we have not seen yet.
                     # Create it's avatar.
+                    # XXX: this needs to be disentangled from pybullet.
+                    # XXX: (adding social vehicles to the vehicle index should not require pybullet to be present)
                     social_vehicle = self._vehicle_index.build_social_vehicle(
                         sim=self,
                         vehicle_state=vehicle,
@@ -565,7 +570,9 @@ class SMARTS:
                         vehicle_type=vehicle.vehicle_type,
                     )
                 # Update the social vehicle avatar to match the vehicle state
-                social_vehicle.control(pose=vehicle.pose, speed=vehicle.speed)
+                social_vehicle.control(
+                    pose=vehicle.pose, speed=vehicle.speed, dt=self._timestep_sec
+                )
 
     def _pybullet_provider_step(self, agent_actions) -> ProviderState:
         self._perform_agent_actions(agent_actions)
@@ -604,6 +611,10 @@ class SMARTS:
     def _nondynamic_provider_step(self, agent_actions) -> ProviderState:
         self._perform_agent_actions(agent_actions)
 
+        self._bullet_client.stepSimulation()
+
+        self._process_collisions()
+
         provider_state = ProviderState()
         nondynamic_agent_ids = {
             agent_id
@@ -618,6 +629,7 @@ class SMARTS:
 
             vehicle = self._vehicle_index.vehicle_by_id(vehicle_id)
             assert isinstance(vehicle.chassis, BoxChassis)
+            vehicle.step(self._elapsed_sim_time)
             provider_state.vehicles.append(
                 VehicleState(
                     vehicle_id=vehicle.id,
@@ -950,7 +962,6 @@ class SMARTS:
             for bubble in self._bubble_manager.bubbles
         ]
 
-        dec_digits = len("{}".format(self._timestep_sec)) - 2
         state = envision_types.State(
             traffic=traffic,
             scenario_id=self.scenario.scenario_hash,
@@ -962,7 +973,7 @@ class SMARTS:
             speed=speed,
             heading=heading,
             lane_ids=lane_ids,
-            frame_time=round(self._elapsed_sim_time + self._total_sim_time, dec_digits),
+            frame_time=self._rounder(self._elapsed_sim_time + self._total_sim_time),
         )
         self._envision.send(state)