+++ smarts/core/traffic_history_provider.py
@@ -18,6 +18,7 @@
 # OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
 # THE SOFTWARE.
 import logging
+import numpy as np
 import sqlite3
 from itertools import cycle
 from typing import NamedTuple, Set
@@ -25,6 +26,7 @@ from typing import NamedTuple, Set
 from .controllers import ActionSpaceType
 from .coordinates import BoundingBox, Heading, Pose
 from .provider import Provider, ProviderState
+from .utils.math import rounder_for_dt
 from .vehicle import VEHICLE_CONFIGS, VehicleState
 
 
@@ -85,7 +87,7 @@ class TrafficHistoryProvider(Provider):
         # Ignore other sim state
         pass
 
-    def _decode_vehicle_type(self, vehicle_type):
+    def _decode_vehicle_type(self, vehicle_type: int) -> str:
         # Options from NGSIM and INTERACTION currently include:
         #  1=motorcycle, 2=auto, 3=truck, 4=pedestrian/bicycle
         if vehicle_type == 1:
@@ -102,13 +104,17 @@ class TrafficHistoryProvider(Provider):
             )
         return "passenger"
 
-    def step(self, provider_actions, dt, elapsed_sim_time) -> ProviderState:
+    def step(
+        self, provider_actions, dt: float, elapsed_sim_time: float
+    ) -> ProviderState:
         if not self._histories:
             return ProviderState(vehicles=[])
         vehicles = []
         vehicle_ids = set()
-        history_time = self._start_time_offset + elapsed_sim_time
-        rows = self._histories.vehicles_active_between(history_time - dt, history_time)
+        rounder = rounder_for_dt(dt)
+        history_time = rounder(self._start_time_offset + elapsed_sim_time)
+        prev_time = rounder(history_time - dt)
+        rows = self._histories.vehicles_active_between(prev_time, history_time)
         for hr in rows:
             v_id = str(hr.vehicle_id)
             if v_id in vehicle_ids or v_id in self._replaced_vehicle_ids:
@@ -116,18 +122,13 @@ class TrafficHistoryProvider(Provider):
             vehicle_ids.add(v_id)
             vehicle_type = self._decode_vehicle_type(hr.vehicle_type)
             default_dims = VEHICLE_CONFIGS[vehicle_type].dimensions
+            pos_x = hr.position_x + self._map_location_offset[0]
+            pos_y = hr.position_y + self._map_location_offset[1]
             vehicles.append(
                 VehicleState(
                     vehicle_id=self._vehicle_id_prefix + v_id,
                     vehicle_type=vehicle_type,
-                    pose=Pose.from_center(
-                        [
-                            hr.position_x + self._map_location_offset[0],
-                            hr.position_y + self._map_location_offset[1],
-                            0,
-                        ],
-                        Heading(hr.heading_rad),
-                    ),
+                    pose=Pose.from_center((pos_x, pos_y, 0), Heading(hr.heading_rad)),
                     dimensions=BoundingBox(
                         length=hr.vehicle_length
                         if hr.vehicle_length is not None